+++ src/query_and_cache/query.py
@@ -1,12 +1,12 @@
 """Query Class to be extended for use with specific sites"""
 
 import logging
-import os
 from typing import Any, NotRequired, Optional, TypedDict
 
-from bs4 import BeautifulSoup
+import requests
 
-from .cache import retrieve_from_cache, retrieve_or_request, store_in_cache
+from query_and_cache.cacheclass import Cache
+from query_and_cache.parser import Parser
 
 
 class QueryConfig(TypedDict):
@@ -16,9 +16,11 @@ class QueryConfig(TypedDict):
     check_cache: NotRequired[bool]
     api_key: NotRequired[str]
     cache_path: NotRequired[str]
+    parser: NotRequired[Parser]
+    cache: NotRequired[Cache]
 
 
-class Query:
+class Query:  # pylint: disable=too-few-public-methods
     """Query Class to be extended for use with specific sites"""
 
     def __init__(self, url: str, config: Optional[QueryConfig] = None):
@@ -28,73 +30,37 @@ class Query:
         self.auth = config.get("auth", None)
         self.check_cache = config.get("check_cache", True)
         self.api_key = config.get("api_key", None)
-        self.cache_path = config.get("cache_path", "cache")
-        self.service_name = self.__class__.__name__.lstrip("Query").lower()
-        if len(self.service_name) == 0:
-            self.service_name = "query"
-        self.logger = logging.getLogger(f"{self.service_name}")
+        cache_path = config.get("cache_path", "cache")
+        self.parser = config.get("parser", None)
+        self.cache = config.get("cache", Cache(cache_path))
+        service_name = self.__class__.__name__.lstrip("Query").lower()
+        if len(service_name) == 0:
+            service_name = "query"
+        self.logger = logging.getLogger(f"{service_name}")
         self.logger.setLevel(logging.DEBUG)
 
-    def get_full_cache_path(self, filename: str) -> str:
-        """Get the full path to the cache file"""
-        return os.path.join(self.cache_path, self.service_name, filename)
-
-    def retrieve_cache(self, search_string: str) -> bool | Any:
-        """Retrieve query data from cache"""
-        if len("".join(e for e in search_string if e.isalnum())) < 1:
-            self.logger.debug("Invalid search string")
-            return False
-        cache_file_path = self.get_full_cache_path(f"{search_string}.json")
-        return retrieve_from_cache(cache_file_path)
-
-    def store_in_cache(self, search_string: str, data: Any) -> bool:
-        """Store query data in cache"""
-        if len("".join(e for e in search_string if e.isalnum())) < 1:
-            return False
-        try:
-            word = data["word"]
-            if not word or len("".join(e for e in word if e.isalnum())) < 1:
-                return False
-        except KeyError as e:
-            word = search_string
-            self.logger.warning("%s", e)
-        cache_file_path = self.get_full_cache_path(f"{word}.json")
-        return store_in_cache(cache_file_path, data)
-
-    def query(self, search_string: str) -> Any:
-        """Query the site with search_string"""
-        search_string = search_string.lower()
-        if self.check_cache:
-            cached: bool | BeautifulSoup = self.retrieve_cache(search_string)
-            if cached is not False and cached is not True and cached is not None:
-                self.logger.info("Search string (%s) found in cache", search_string)
-                return self.parse_soup(cached)
-            self.logger.info("Search string (%s) not found in cache", search_string)
+    def query(self, query_string: str) -> Any:
+        """Query the site with query_string"""
+        query_string = query_string.lower()
+        if self.check_cache and self.cache:
+            cached: Any | None = self.cache.retrieve(query_string)
+            if cached is not None:
+                self.logger.info("Search string (%s) found in cache", query_string)
+                return cached
+            self.logger.info("Search string (%s) not found in cache", query_string)
         else:
             self.logger.info("Skipping Cache as requested")
-        url = self.url.format(search_string=search_string, api_key=self.api_key)
+        url = self.url.format(search_string=query_string, api_key=self.api_key)
         self.logger.debug("querying %s", url)
-        webpage = retrieve_or_request(
-            url, self.get_full_cache_path(f"{search_string}.html")
-        )
-        if webpage is False or webpage is True or webpage is None:
-            self.logger.error("Error retrieving webpage")
-            return False
-        soup = BeautifulSoup(webpage, features="html.parser")
-        results = self.parse_soup(soup)
-        self.logger.debug("Results received from soup parser")
-        if "word" not in results or results["word"] is None:
-            results["word"] = search_string
-        try:
-            self.store_in_cache(search_string, results)
-        except KeyError as e:
-            self.logger.error("Error storing in cache: results does not have %s key", e)
-        self.logger.debug("Returning query results")
-        return results
-
-    def parse_soup(self, soup: BeautifulSoup) -> Any:
-        """Parse the webpage to find the desired information
-
-        This method should be overridden and designed to properly
-        handle data from the given website"""
-        raise NotImplementedError
+        result = requests.get(url, timeout=5)
+        if result.status_code != 200:
+            self.logger.error("Query of %s failed (%s)", url, result.status_code)
+            return None
+        if self.parser:
+            self.logger.debug("Parsing data")
+            result = self.parser.parse(result)
+        else:
+            self.logger.debug("No parser found, returning raw data")
+        if self.check_cache and self.cache:
+            self.cache.store(query_string, result)
+        return result