+++ tests/test_query_class.py
@@ -30,7 +30,6 @@ class Test(unittest.TestCase):
         self.assertEqual(query.auth, "auth_type")
         self.assertEqual(query.check_cache, False)
         self.assertEqual(query.api_key, "fakekey")
-        self.assertEqual(query.cache_path, "new_cache")
 
     def test_query_has_cache_on_by_default(self) -> None:
         """
@@ -38,7 +37,3 @@ class Test(unittest.TestCase):
         """
         query = Query("localhost")
         self.assertEqual(query.check_cache, True)
-
-
-if __name__ == "__main__":
-    unittest.main()